Verificaciones PDF.

+++ graphs.py
@@ -118,28 +118,31 @@ class Graphs():
     def donwload_graphs(self):
        
         # Ajustar los límites para eliminar espacio en blanco
-        self.figure_bar_l.gca().set_ylim(0, (max(self.defl_l_data)+100))
-        self.figure_bar_r.gca().set_ylim(0, (max(self.defl_r_data)+100))  # Ajustar límites en el eje y según tu necesidad
+        if(self.defl_l_data==[] or self.defl_r_data==[]):
+            return
+        else:
+            self.figure_bar_l.gca().set_ylim(0, (max(self.defl_l_data)+100))
+            self.figure_bar_r.gca().set_ylim(0, (max(self.defl_r_data)+100))  # Ajustar límites en el eje y según tu necesidad
 
-        self.figure_bar_l.savefig('figure_bar_l.png', bbox_inches='tight')
-        self.figure_bar_r.savefig('figure_bar_r.png', bbox_inches='tight')
-    
-        # Crear un nuevo PDF con ambas figuras
-        output_pdf = 'defl_individuales.pdf'
+            self.figure_bar_l.savefig('figure_bar_l.png', bbox_inches='tight')
+            self.figure_bar_r.savefig('figure_bar_r.png', bbox_inches='tight')
         
-        c = canvas.Canvas(output_pdf, pagesize=A4)
+            # Crear un nuevo PDF con ambas figuras
+            output_pdf = 'defl_individuales.pdf'
+            
+            c = canvas.Canvas(output_pdf, pagesize=A4)
 
-        # Agregar la primera figura en la posición deseada
-        c.drawImage('figure_bar_l.png', 10, 0)
+            # Agregar la primera figura en la posición deseada
+            c.drawImage('figure_bar_l.png', 10, 0)
 
-        # Agregar la segunda figura debajo de la primera
-        c.drawImage('figure_bar_r.png', 10, 500)
+            # Agregar la segunda figura debajo de la primera
+            c.drawImage('figure_bar_r.png', 10, 500)
 
-        # Guardar el contenido en el PDF
-        c.save()
+            # Guardar el contenido en el PDF
+            c.save()
 
-        os.remove('figure_bar_l.png')
-        os.remove('figure_bar_r.png')
+            os.remove('figure_bar_l.png')
+            os.remove('figure_bar_r.png')
